+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+#v1.2
+* added feature, which allows to fetch all the features/fields from the data directory.
+
 #v1.0.2
-* fixed a bug where if you extend but not mirror it would crash
+* fixed a bug where if you extend but not mirror it would crash.
 
 #v1.0.1
 * changed the way you set the extension. Now only need to input the number of cells to extend by. `0` corresponds to no extension. No longer a need for boolean input.